+++ src/solc_0.8/common/interfaces/IGameToken.sol
@@ -18,7 +18,7 @@ interface IGameToken {
         address to,
         Update calldata creation,
         address editor,
-        uint64 subId
+        uint48 subId
     ) external returns (uint256 id);
 
     function burn(uint256 gameId) external;
